+++ src/components/user/uc/users.uc.js
@@ -1,7 +1,7 @@
 const { ObjectId } = require('mongoose').Types;
-const { GeneralError, NotFound, Forbidden, BadRequest } = require('../../../errors');
+const { GeneralError, NotFound } = require('../../../errors');
 const { userRepo, accountRepo, trashbinRepo } = require('../repo/index');
-const { equal: equalIds } = require('../../../helper/compare').ObjectId;
+const restrictToSameSchool = require('../../../utils/restrictToSameSchool');
 
 const getUserData = async (id, app) => {
 	const data = {};
@@ -44,19 +44,6 @@ const replaceUserWithTombstone = async (id, app) => {
 		app
 	);
 	return { success: true };
-};
-
-const restrictToSameSchool = async (id, account, app) => {
-	if (id) {
-		const { schoolId: currentUserSchoolId } = await userRepo.getUser(account.userId, app);
-		const { schoolId: requestedUserSchoolId } = await userRepo.getUser(id, app);
-
-		if (!equalIds(currentUserSchoolId, requestedUserSchoolId)) {
-			throw new Forbidden('You have no access.');
-		}
-		return true;
-	}
-	throw new BadRequest('The request query should include a valid userId');
 };
 
 const deleteUserUC = async (id, { account, app }) => {
